requirements.txt
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
greenlet==2.0.1
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
playwright==1.31.1
pluggy==1.0.0
pyee==9.0.4
pytest==7.2.2
pytest-base-url==2.0.0
pytest-playwright==0.3.2
python-dotenv==1.0.0
python-slugify==6.1.2
requests==2.28.2
text-unidecode==1.3
typing_extensions==4.5.0
urllib3==1.26.15
